--- a/docs/Fields.md
+++ b/docs/Fields.md
@@ -94,10 +94,12 @@ Ideal for embedded arrays of objects, e.g. `tags` and `backlinks` in the followi
   ],
   backlinks: [
         {
+            uuid: '34fdf393-f449-4b04-a423-38ad02ae159e',
             date: '2012-08-10T00:00:00.000Z',
             url: 'http://example.com/foo/bar.html',
         },
         {
+            uuid: 'd907743a-253d-4ec1-8329-404d4c5e6cf1',
             date: '2012-08-14T00:00:00.000Z',
             url: 'https://blog.johndoe.com/2012/08/12/foobar.html',
         }
@@ -108,9 +110,11 @@ Ideal for embedded arrays of objects, e.g. `tags` and `backlinks` in the followi
 The child must be an iterator component (like `<Datagrid>` or `<SingleFieldList>`).
 
 Here is how to display all the backlinks of the current post as a `<Datagrid>`
+Optional prop fieldKey can be pass to ArrayField to use as Key
+Which reduces time and memory to generate key
 
 ```jsx
-<ArrayField source="backlinks">
+<ArrayField source="backlinks" fieldKey="uuid">
     <Datagrid>
         <DateField source="date" />
         <UrlField source="url" />