**Is your feature request related to a problem? Please describe.**
using JSON.stringify results with a very large key. Passing a fieldKey when available reduces the overhead