@fzaninotto 2020-02-21
Hi, and thanks for your patch. 

Did you measure the actual overhead of the JSON stringification? Can you provide metrics showing the improvement due to your change?