@fzaninotto 2020-02-25
Thanks, but I meant real-life usage. 

I have no doubt that skipping stringification is faster than using it. But my point is: on an average scenario of ArrayField usage, what's the overhead of the JSON stringification as compared to all the other JS execution? 

My intuition is that, unless you're using ArrayField with a very large array of very large objects, the cost of stringifying JSON isn't significant when compared to the cost of rendering. I may be wrong of course, but before we fix a potential bug, can you show me the actual bug?