@Hemant-yadav 2020-02-26
Hi,

With small data the difference is insignificant compared to rendering but with 50 records total of ~643KB data here are the results.

I have created a [sandbox](https://codesandbox.io/s/funny-frog-yqgz0) with the exact code I used.

**Without FieldKey**
<img width="1280" alt="withoutFieldKey" src="https://user-images.githubusercontent.com/5657104/75361660-a8b9ff80-58dd-11ea-9fc2-e633d845ed40.png">

**With FieldKey**
<img width="1280" alt="withFieldKey" src="https://user-images.githubusercontent.com/5657104/75361712-bff8ed00-58dd-11ea-9313-ee2f05e5d9f4.png">

The time difference between rendering the ArrayField was [~10ms, ~30ms]and total rendering was also reduced by [~50ms, ~100ms]. The memory size is also reduced by ~2MB.