@fzaninotto 2020-02-27
Thanks for taking the time. That's convincing, so we'll consider merging it.

But before that, you need to change the base branch to `next` (as it's not a bug fix), and add documentation for the new prop.